+++ app/src/main/AndroidManifest.xml
@@ -13,7 +13,7 @@
         android:theme="@style/Theme.KillerGram_AndroidV1"
         tools:targetApi="31">
         <activity
-            android:name=".MainActivity"
+            android:name=".feature.MainActivity"
             android:windowSoftInputMode="adjustResize"
             android:exported="true">
             <intent-filter>
