+++ app/src/main/java/com/example/killergram_android_v1/feature/MainActivity.kt
@@ -1,0 +1,13 @@
+package com.example.killergram_android_v1.feature
+
+import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
+import com.example.killergram_android_v1.R
+
+class MainActivity : AppCompatActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_main)
+
+    }
+}
