+++ app/src/main/java/com/example/killergram_android_v1/feature/login/LoginFragment.kt
@@ -1,15 +1,10 @@
-package com.example.killergram_android_v1
+package com.example.killergram_android_v1.feature.login
 
 import android.os.Bundle
-import android.text.Editable
-import android.text.TextWatcher
-import android.util.Log
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.widget.Button
-import androidx.navigation.fragment.findNavController
 
 
 class LoginFragment : Fragment() {
