+++ app/src/main/java/com/example/killergram_android_v1/feature/signup/AuthCodeFragment.kt
@@ -1,0 +1,22 @@
+package com.example.killergram_android_v1.feature.signup
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import com.example.killergram_android_v1.R
+
+
+class AuthCodeFragment : Fragment() {
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_email_verification, container, false)
+    }
+
+
+}
