+++ app/src/main/java/com/example/killergram_android_v1/feature/signup/SetPasswordFragment.kt
@@ -1,4 +1,4 @@
-package com.example.killergram_android_v1.signup
+package com.example.killergram_android_v1.feature.signup
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -14,10 +14,10 @@ private const val ARG_PARAM2 = "param2"
 
 /**
  * A simple [Fragment] subclass.
- * Use the [AuthCodeFragment.newInstance] factory method to
+ * Use the [SetPasswordFragment.newInstance] factory method to
  * create an instance of this fragment.
  */
-class AuthCodeFragment : Fragment() {
+class SetPasswordFragment : Fragment() {
     // TODO: Rename and change types of parameters
     private var param1: String? = null
     private var param2: String? = null
@@ -35,7 +35,7 @@ class AuthCodeFragment : Fragment() {
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_email_verification, container, false)
+        return inflater.inflate(R.layout.fragment_set_password, container, false)
     }
 
     companion object {
@@ -45,12 +45,12 @@ class AuthCodeFragment : Fragment() {
          *
          * @param param1 Parameter 1.
          * @param param2 Parameter 2.
-         * @return A new instance of fragment EmailVerificationFragment.
+         * @return A new instance of fragment SetPasswordFragment.
          */
         // TODO: Rename and change types and number of parameters
         @JvmStatic
         fun newInstance(param1: String, param2: String) =
-            AuthCodeFragment().apply {
+            SetPasswordFragment().apply {
                 arguments = Bundle().apply {
                     putString(ARG_PARAM1, param1)
                     putString(ARG_PARAM2, param2)
