+++ app/src/main/res/layout/activity_main.xml
@@ -7,7 +7,7 @@
         android:id="@+id/main"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".MainActivity">
+        tools:context=".feature.MainActivity">
 
         <FrameLayout
             android:layout_width="match_parent"
