+++ app/src/main/res/layout/fragment_sign_up.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:background="@color/back_black"
-    tools:context=".signup.EmailVerificationFragment">
+    tools:context=".feature.signup.EmailVerificationFragment">
 
     <androidx.constraintlayout.widget.Guideline
         android:id="@+id/glTop"
